pyproject.toml
[project]
name = "the_iron_utility"
version = "0.1.0"
description = "Real-time squat form analysis using webcam video."
authors = [
    { name = "GeminiTrainer", email = "gemini@example.com" }
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.2.0",
    "hydra-core>=1.3.0",
    "mlflow>=2.0.0",
    "uv>=0.1.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "Apache-2.0" }

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "D", "UP", "ANN", "ASYNC", "B", "C4", "DTZ", "ERA", "ISC", "ICN", "PIE", "PT", "Q", "RSE", "RET", "SIM", "TID", "TCH", "INT", "ARG", "NPY", "PERF", "FURB"]
ignore = [
    "D100", # Missing docstrings in public modules
    "D104", # Missing docstrings in public package
    "D105", # Missing docstrings in public methods
    "D107", # Missing docstrings in __init__
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"] # Allow unused imports in __init__.py

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["src"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true
implicit_reexport = false
strict_optional = true
show_error_codes = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--strict-markers --import-mode=append"
testpaths = ["tests"]
python_files = "test_*.py"
log_cli_level = "INFO"
console_output_style = "classic"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
